requirements.txt
pyocr>=0.4.7
Wand>=0.4.4
tesserocr>=2.1.3
google-cloud-vision>=0.31.1
